app/requirements.txt
kserve
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.0.3
scikit-learn==1.3.2
mlflow==2.19.0
boto3
joblib
python-dotenv
pydantic